+++ src/commands/validate-tasks.ts
@@ -28,6 +28,7 @@ export class ValidateTasksCommand extends BaseCliCommand<IPerformTasksCommandArg
         command.option('--parameters [parameters]', 'parameters used when creating build tasks from tasks file');
         command.option('--organization-state-object [organization-state-object]', 'key for object used to load read-only organization state');
         command.option('--organization-state-bucket-name [organization-state-bucket-name]', 'name of the bucket that contains the read-only organization state');
+        command.option('--debug-templating [debug-templating]', 'when set to true the output of text templating processes will be stored on disk', false);
 
         super.addOptions(command);
     }
@@ -41,10 +42,10 @@ export class ValidateTasksCommand extends BaseCliCommand<IPerformTasksCommandArg
         Validator.validatePositiveInteger(command.failedTasksTolerance, 'failedTasksTolerance');
 
         command.parsedParameters = this.parseCfnParameters(command.parameters);
-        const config = new BuildConfiguration(tasksFile, command.parsedParameters);
+        const config = new BuildConfiguration(tasksFile, command.parsedParameters, command.TemplatingContext);
 
         await config.fixateOrganizationFile(command);
         const validationTasks = config.enumValidationTasks(command);
-        await BuildRunner.RunValidationTasks(validationTasks, command.verbose === true , command.maxConcurrentTasks, command.failedTasksTolerance);
+        await BuildRunner.RunValidationTasks(validationTasks, command.verbose === true, command.maxConcurrentTasks, command.failedTasksTolerance);
     }
 }