+++ CHANGELOG.md
@@ -2,7 +2,18 @@
 
 All notable changes to aws organization formation will be documented in this file.
 
-**version 0.9.16**
+**unreleased**
+
+- fix: allow templates without version attribute
+- feat: support `!Ref AWS::Partition` in tasks file
+- feat: allow yaml anchors to be declared in top level `Definitions` attribute
+- feat: reduce 'not in update create completed' failures when creating new account (due to rate limiting)
+- feat: TextTemplating support for organization.yml and task files
+- feat: added possibility to mix TextTemplating and bindings/expressions
+- feat: stack tags
+- fix: org-formation init to honor `--cross-account-role-name`
+
+- **version 0.9.16**
 
 - fix: missing role in build account when running init-pipeline